All crud methods working

+++ botigaCuadres/src/main/java/com/itacademy/botigaCuadres/repository/PaintingRepository.java
@@ -8,7 +8,7 @@ import com.itacademy.botigaCuadres.dto.PaintingResponseDto;
 import com.itacademy.botigaCuadres.dto.ShopResponseDto;
 
 @Repository
-public interface PaintingRepository extends JpaRepository<PaintingResponseDto, Long> {
+public interface PaintingRepository extends JpaRepository<PaintingResponseDto, Integer> {
 	
 	Iterable<PaintingResponseDto> findByShop(ShopResponseDto shop);
